requirements.txt
wtpsplit
python-dotenv
ollama
google-generativeai
openai
pydantic
pydantic
python-dotenv
torch